Fixed CCE in compareTo.
git-svn-id: https://svn.neo4j.org/components/neo/trunk@960 0b971d98-bb2f-0410-8247-b05b2b5feb2a

+++ community/src/java/org/neo4j/impl/core/NodeImpl.java
@@ -464,7 +464,7 @@ class NodeImpl extends NeoPrimitive implements Node, Comparable<Node>
 	 */
 	public int compareTo( Node n )
 	{
-		int ourId = (int) this.getId(), theirId = (int) n.getId();
+		long ourId = (long) this.getId(), theirId = (long) n.getId();
 		
 		if ( ourId < theirId )
 		{
